:recycle: refactor(pr): Code changes for PR
