+++ src/@types/Timeframe.ts
@@ -1,10 +1,3 @@
-export enum Timeframes {
-  MAX = `max`,
-  ONE_YEAR = `12m`,
-  YEAR_TO_DATE = `ytd`,
-  ONE_MONTH = `1m`,
-}
-
 export enum TimeframesData {
   MAX = `max`,
   ONE_YEAR = `twelveMonths`,
