+++ src/styles/Font.ts
@@ -1,0 +1,4 @@
+import { Inter } from 'next/font/google'
+
+/*eslint @typescript-eslint/quotes: off*/
+export default Inter({ subsets: ['latin'] })
