+++ src/styles/FontSizes.ts
@@ -1,0 +1,14 @@
+const fontSizes = {
+  xs: `10px`,
+  sm: `12px`,
+  md: `14px`,
+  lg: `16px`,
+  xl: `18px`,
+  '2xl': `22px`,
+  '3xl': `28px`,
+  '4xl': `36px`,
+  '5xl': `48px`,
+  '6xl': `64px`,
+}
+
+export default fontSizes
